tokengen: address compiler warnings (#1633)
Remove redundant casts, address raw type warnings, and unchecked calls
in JRandTest code to compile the add-on with all lint checks and with
warnings as error.

+++ src/com/fasteasytrade/JRandTest/Algo/MT19937Prng.java
@@ -225,7 +225,7 @@ public class MT19937Prng extends Cipher {
 	 * generates a random number on [0,0x7fffffff]-interval
 	 */
 	long genrand_int31() {
-		return (long) (genrand_int32() >>> 1);
+		return genrand_int32() >>> 1;
 	}
 
 	/**
